refactor: ErrorResponse dto 패키지 분리
